Refactor successor generator and make cleverer spooling successor generator

+++ include/LTL/Structures/ProductState.h
@@ -51,7 +51,7 @@ namespace LTL::Structures {
         }
 
         [[nodiscard]] bool markingEqual(const PetriEngine::MarkVal *rhs) const {
-            for (int i = 0; i < buchi_state_idx; ++i) {
+            for (size_t i = 0; i < buchi_state_idx; ++i) {
                 if (marking()[i] != rhs[i]) {
                     return false;
                 }
